# swagger.yaml
...
swagger: "2.0"
basePath: "/api/v0.2"
info:
  title: LEXIS Portal API
  description: The LEXIS Portal API serves as the API proxy to all the interfaces that interact with the parts involved in LEXIS.
  version: "0.2.2021-12-15"
  contact:
    email: diego@cyclops-labs.io
  license:
    name: Apache 2.0
    url: 'http://www.apache.org/licenses/LICENSE-2.0.html'
...
securityDefinitions:
  APIKeyHeader:
    type: apiKey
    in: header
    name: X-API-KEY
  APIKeyParam:
    type: apiKey
    in: query
    name: api_key
  Keycloak:
    type: oauth2
    flow:  accessCode
    # Need to get clarity on whether these are necessary here or provided with a config file
    authorizationUrl: 'https://138.246.232.245:6443/auth'
    tokenUrl: 'https://138.246.232.245:6443/auth'
security:
  - APIKeyHeader: []
  - APIKeyParam: []
  - Keycloak: []
paths:
  /accounting/{id}/usage:
    get:
      tags:
        - usageManagement
      consumes:
        - application/json
      produces:
        - application/json
      summary: Get account usage for specific project
      operationId: getUsage
      description: get account usage of project with given id
      responses:
        '200':
          description: usage returned
          schema:
            $ref: "#/definitions/Usage"
        '401':
          description: Unauthorized
          schema:
            $ref: "#/definitions/ErrorResponse"
        '403':
          description: Forbidden
          schema:
            $ref: "#/definitions/ErrorResponse"
        '404':
          description: project not found
          schema:
            $ref: "#/definitions/ErrorResponse"
        '500':
          description: unexpected error
          schema:
            $ref: "#/definitions/ErrorResponse"
      parameters:
        - in: path
          name: id
          type: string
          format: uuid
          required: true
          description: Id of project to be obtained
        - name: from
          in: query
          description: Datetime from which to get the usage report
          type: string
          format: datetime
        - name: to
          in: query
          description: Datetime until which to get the usage report
          type: string
          format: datetime
  /accounting/{id}/available:
    get:
      tags:
        - usageManagement
      produces:
        - application/json
      summary: Credit status of the account with the provided id
      operationId: getCredit
      responses:
        '200':
          description: Credit status of the account with the provided id
          schema:
            $ref: "#/definitions/CreditStatus"
        '401':
          description: Unauthorized
          schema:
            $ref: "#/definitions/ErrorResponse"
        '403':
          description: Forbidden
          schema:
            $ref: "#/definitions/ErrorResponse"
        '404':
          description: The account with the provided id doesn't exist
          schema:
            $ref: "#/definitions/ErrorResponse"
        '500':
          description: Something unexpected happend, error raised
          schema:
            $ref: "#/definitions/ErrorResponse"
      parameters:
        - name: id
          in: path
          type: string
          format: uuid
          required: true
          description: Id of the account to be checked
  /accounting/{id}/history:
    get:
      tags:
        - usageManagement
      produces:
        - application/json
      summary: Credit history of the customer with id
      operationId: getHistory
      responses:
        '200':
          description: Credit status history of the account with the provided id
          schema:
            $ref: "#/definitions/CreditHistory"
        '401':
          description: Unauthorized
          schema:
            $ref: "#/definitions/ErrorResponse"
        '403':
          description: Forbidden
          schema:
            $ref: "#/definitions/ErrorResponse"
        '404':
          description: The endpoint provided doesn't exist
          schema:
            $ref: "#/definitions/ErrorResponse"
        '500':
          description: Something unexpected happend, error raised
          schema:
            $ref: "#/definitions/ErrorResponse"
      parameters:
        - name: id
          in: path
          type: string
          format: uuid
          required: true
          description: Id of the account to get the history
        - name: filterSystem
          in: query
          type: boolean
          description: Boolean variable to control if the system consumptions have to be listed or not
        - name: medium
          in: query
          description: Medium (cash/credit) to be used as filter
          type: string
          enum:
          - credit
          - cash
  /accounting/{id}/manage/{medium}/decrease/{amount}:
    post:
      tags:
        - usageManagement
      consumes:
        - application/json
      produces:
        - application/json
      summary: Insert a new reseller in the system.
      operationId: decreaseCredit
      responses:
        '200':
          description: Credit status of the account with the provided id
          schema:
            $ref: "#/definitions/CreditStatus"
        '401':
          description: Unauthorized
          schema:
            $ref: "#/definitions/ErrorResponse"
        '403':
          description: Forbidden
          schema:
            $ref: "#/definitions/ErrorResponse"
        '404':
          description: The account with the id provided doesn't exist
          schema:
            $ref: "#/definitions/ErrorResponse"
        '500':
          description: Something unexpected happend, error raised
          schema:
            $ref: "#/definitions/ErrorResponse"
      parameters:
        - name: id
          in: path
          type: string
          format: uuid
          required: true
          description: Id of the account to be checked
        - name: amount
          in: path
          description: Amount to be decreased
          required: true
          type: number
          format: double
        - name: medium
          in: path
          description: Medium (cash/credit) to be used in the accounting
          required: true
          type: string
          enum:
          - credit
          - cash
  /accounting/{id}/manage/{medium}/increase/{amount}:
    post:
      tags:
        - usageManagement
      consumes:
        - application/json
      produces:
        - application/json
      summary: Insert a new reseller in the system.
      operationId: increaseCredit
      responses:
        '200':
          description: Credit status of the account with the provided id
          schema:
            $ref: "#/definitions/CreditStatus"
        '401':
          description: Unauthorized
          schema:
            $ref: "#/definitions/ErrorResponse"
        '403':
          description: Forbidden
          schema:
            $ref: "#/definitions/ErrorResponse"
        '404':
          description: The account with the id provided doesn't exist
          schema:
            $ref: "#/definitions/ErrorResponse"
        '500':
          description: Something unexpected happend, error raised
          schema:
            $ref: "#/definitions/ErrorResponse"
      parameters:
        - name: id
          in: path
          type: string
          format: uuid
          required: true
          description: Id of the account to be checked
        - name: amount
          in: path
          description: Amount to be inccreased
          required: true
          type: number
          format: double
        - name: medium
          in: path
          description: Medium (cash/credit) to be used in the accounting
          required: true
          type: string
          enum:
          - credit
          - cash
  /accounting/{id}/manage/{medium}/consume/{amount}:
    post:
      tags:
        - usageManagement
      consumes:
        - application/json
      produces:
        - application/json
      summary: Adds a consumption to the system
      operationId: addConsumption
      responses:
        '200':
          description: Credit status of the account with the provided id
          schema:
            $ref: "#/definitions/CreditStatus"
        '401':
          description: Unauthorized
          schema:
            $ref: "#/definitions/ErrorResponse"
        '403':
          description: Forbidden
          schema:
            $ref: "#/definitions/ErrorResponse"
        '404':
          description: The account with the id provided doesn't exist
          schema:
            $ref: "#/definitions/ErrorResponse"
        '500':
          description: Something unexpected happend, error raised
          schema:
            $ref: "#/definitions/ErrorResponse"
      parameters:
        - name: id
          in: path
          type: string
          format: uuid
          required: true
          description: Id of the account to be checked
        - name: amount
          in: path
          description: Amount to be decreased
          required: true
          type: number
          format: double
        - name: medium
          in: path
          description: Medium (cash/credit) to be used in the accounting
          required: true
          type: string
          enum:
          - credit
          - cash
...
  /project/{id}:
    get:
      tags:
        - userOrgManagement
      consumes:
        - application/json
      produces:
        - application/json
      summary: Get specific project
      operationId: getProject
      description: get project with given id
      responses:
        '200':
          description: project returned
          schema:
            $ref: "#/definitions/Project"
        '404':
          description: project with not found
          schema:
            $ref: "#/definitions/ErrorResponse"
        '401':
          description: Unauthorized
          schema:
            $ref: "#/definitions/ErrorResponse"
        '403':
          description: Forbidden
          schema:
            $ref: "#/definitions/ErrorResponse"
        '500':
          description: unexpected error
          schema:
            $ref: "#/definitions/ErrorResponse"
      parameters:
        - in: path
          name: id
          type: string
          format: uuid
          required: true
          description: Id of project to be obtained
    put:
      tags:
        - userOrgManagement
      consumes:
        - application/json
      produces:
        - application/json
      summary: Update specific project
      operationId: updateProject
      description: Update project with given id
      responses:
        '200':
          description: updated project
          schema:
            $ref: "#/definitions/Project"
        '404':
          description: project with not found
          schema:
            $ref: "#/definitions/ErrorResponse"
        '422':
          description: Shortname duplicated
          schema:
            $ref: "#/definitions/ErrorResponse"
        '401':
          description: Unauthorized
          schema:
            $ref: "#/definitions/ErrorResponse"
        '403':
          description: Forbidden
          schema:
            $ref: "#/definitions/ErrorResponse"
        '500':
          description: unexpected error
          schema:
            $ref: "#/definitions/ErrorResponse"
      parameters:
        - in: path
          name: id
          type: string
          format: uuid
          required: true
          description: Id of project to be obtained
        - in: body
          name: project
          required: true
          description: updated project data to be added
          schema:
            $ref: "#/definitions/Project"
    delete:
      tags:
        - userOrgManagement
      consumes:
        - application/json
      produces:
        - application/json
      summary: Delete specific project
      operationId: deleteProject
      description: Delete project with given id
      responses:
        '200':
          description: deleted project
          schema:
            $ref: "#/definitions/DeletedResponse"
        '404':
          description: project with not found
          schema:
            $ref: "#/definitions/ErrorResponse"
        '401':
          description: Unauthorized
          schema:
            $ref: "#/definitions/ErrorResponse"
        '403':
          description: Forbidden
          schema:
            $ref: "#/definitions/ErrorResponse"
        '500':
          description: unexpected error
          schema:
            $ref: "#/definitions/ErrorResponse"
      parameters:
        - in: path
          name: id
          type: string
          format: uuid
          required: true
          description: Id of project to be obtained
...
  /hpc/resource:
    get:
      tags:
        - userOrgManagement
      produces:
        - application/json
      summary: list HPCResource
      operationId: listHPCResources
      description: lists all HPCResources
      responses:
        '200':
          description: list of HPCResources
          schema:
            type: array
            items:
              $ref: "#/definitions/HPCResource"
        '401':
          description: Unauthorized
          schema:
            $ref: "#/definitions/ErrorResponse"
        '403':
          description: Forbidden
          schema:
            $ref: "#/definitions/ErrorResponse"
        '500':
          description: unexpected error
          schema:
            $ref: "#/definitions/ErrorResponse"
      parameters:
        - in: query
          name: scope
          description: organization scope switch
          type: string
          enum:
          - ALL
          - OWN
          - ALLOWED
          default: ALL
    post:
      tags:
        - userOrgManagement
      consumes:
        - application/json
      produces:
        - application/json
      summary: create HPCResource
      operationId: createHPCResource
      description: Creates a new HPCResource
      responses:
        '201':
          description: HPCResource created
          schema:
            $ref: "#/definitions/ItemCreatedResponse"
        '409':
          description: an existing item already exists
          schema:
            $ref: "#/definitions/ErrorResponse"
        '401':
          description: Unauthorized
          schema:
            $ref: "#/definitions/ErrorResponse"
        '403':
          description: Forbidden
          schema:
            $ref: "#/definitions/ErrorResponse"
        '500':
          description: unexpected error
          schema:
            $ref: "#/definitions/ErrorResponse"
      parameters:
        - in: body
          name: HPCResource
          description: HPCResource to be added
          schema:
            $ref: '#/definitions/HPCResource'
...
definitions:
...
  ItemCreatedResponse:
    type: object
    required:
      - Link
    properties:
      ID:
        type: string
      Link:
        type: string
  DeletedResponse:
    type: object
    properties:
      ID:
        type: string
      Message:
        type: string
...
  ErrorResponse:
    type: object
    required:
      - errorString
    properties:
      errorString:
        type: string
...
  StringArray:
    x-go-type:
      import:
        package: "github.com/lib/pq"
      type: StringArray
  Metadata:
    type: object
    x-go-type:
      import:
        package: "gitlab.com/cyclops-utilities/datamodels"
      type: JSONdb

#HEAppE
...
  Project:
    type: object
    properties:
      ProjectID:
        type: string
        format: uuid
      ProjectName:
        type: string
      ProjectShortName:
        type: string
      ProjectDescription:
        type: string
      ProjectCreationTime:
        type: string
        format: date-time
      ProjectCreatedBy:
        type: string
        format: uuid
      LinkedOrganization:
        type: string
        format: uuid
      AllowedOrganizations:
        $ref: '#/definitions/StringArray'
      ProjectStatus:
        type: string
        enum:
        - PENDING
        - ACTIVE
        - DISABLED
        - TERMINATED
      ProjectContactPerson:
        type: string
        format: uuid
      ProjectStartDate:
        type: string
        format: date-time
      ProjectTerminationDate:
        type: string
        format: date-time
      ProjectMaxPrice:
        type: number
        format: double
        default: 0.0
        x-nullable: true
      NormCoreHours:
        type: integer
        default: 0
        x-nullable: true
      ProjectContactEmail:
        type: string
        format: email
      ProjectDomain:
        type: string
  HPCResource:
    type: object
    properties:
      HPCResourceID:
        type: string
      AssociatedHPCProject:
        type: string
      AssociatedLEXISProject:
        type: string
        format: uuid
      ApprovalStatus:
        type: string
        enum:
        - ACCEPTED
        - REJECTED
        - PENDING
      CloudNetworkName:
        type: string
      ProjectNetworkName:
        type: string
      HEAppEEndpoint:
        type: string
      HPCProvider:
        type: string
        enum:
        - IT4I
        - LRZ
        - ICHEC
      OpenStackEndpoint:
        type: string
      OpenStackProjectID:
        type: string
      ResourceType:
        type: string
        enum:
        - CLOUD
        - HPC
        - SMARTGW
      TermsConsent:
        type: boolean


#Cyclops
  HPCUsage:
    type: object
    properties:
      HPCProjectID:
        type: string
      HPCProjectName:
        type: string
      AccountingData:
        type: array
        items:
          $ref: '#/definitions/Metadata'
  Usage:
    type: object
    properties:
      ProjectID:
        type: string
        format: uuid
      ProjectName:
        type: string
      TimeFrom:
        type: string
        format: datetime
      TimeTo:
        type: string
        format: datetime
      HPCProjects:
        type: array
        items:
          $ref: '#/definitions/HPCUsage'
  CreditHistory:
    type: object
    properties:
      AccountID:
        type: string
      Events:
        type: array
        items:
          $ref: '#/definitions/Event'
  CreditStatus:
    type: object
    properties:
      AccountID:
        type: string
      AvailableCredit:
        type: number
        format: double
      AvailableCash:
        type: number
        format: double
      LastUpdate:
        type: string
        format: datetime
...